Revert "merge main"

+++ app/cruds/slot.py
@@ -1,17 +1,18 @@
 import datetime
-
-from fastapi import HTTPException, status
-from sqlalchemy import update
-from sqlalchemy.future import select
+from fastapi import FastAPI, HTTPException, status
+from app.models.models import User
+from app.schemas.slot import SlotRequest, SlotUpdate
+from app.models.models import Slot, Bidder, Task, Bid
 from sqlalchemy.orm import Session
-
 from app.cruds.response import (
-    slot_response,
     slots_response,
+    slot_response,
     user_response,
+    bidder_response,
 )
-from app.models.models import GroupUser, Slot, Task, User
-from app.schemas.slot import SlotCreate
+from sqlalchemy.future import select
+from app.cruds.bidder import bidder_get
+from sqlalchemy import delete, update
 
 
 def all(db: Session):
@@ -22,7 +23,9 @@ def all(db: Session):
 
 def slot_finished(db: Session):
     item = (
-        db.execute(select(Slot).filter(Slot.end_time < datetime.datetime.now()))
+        db.execute(
+            select(Slot).filter(Slot.end_time < datetime.datetime.now())
+        )
         .scalars()
         .all()
     )
@@ -35,7 +38,7 @@ def get(name: str, db: Session):
     return respone_slot
 
 
-def post(slot: SlotCreate, db: Session, user: User):
+def post(slot: SlotRequest, db: Session, user: User):
     task = db.get(Task, slot.task_id)
     if not task:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST)
@@ -61,10 +64,10 @@ def post(slot: SlotCreate, db: Session, user: User):
     db.add(slot)
     db.commit()
     db.refresh(slot)
-    return slot
+    return slot_response(slot)
 
 
-def patch(request: SlotCreate, slot_id: str, db: Session):
+def patch(request: SlotUpdate, slot_id: str, db: Session):
     slot = db.get(Slot, slot_id)
     if not slot:
         raise HTTPException(
@@ -100,57 +103,73 @@ def patch(request: SlotCreate, slot_id: str, db: Session):
     )
     slot.task = task
     db.commit()
-    db.refresh(slot)
-    return slot
+    return slot_response(slot)
 
 
-def assign(slot_id: str, user_id: str, db: Session):
+def cancel(slot_id: str, user: User, premire_point: int, db: Session):
     slot = db.get(Slot, slot_id)
-    user = db.get(User, user_id)
-    if not slot or not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
-    if slot.end_time < datetime.datetime.now():
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+    bid = slot.bid
+    if bid is None:
+        raise HTTPException(status_code=status.HTTP_405_METHOD_NOT_ALLOWED)
+    bidder = db.scalars(
+        select(Bidder)
+        .filter(Bidder.bid_id == bid.id, Bidder.user_id == user.id)
+        .limit(1)
+    ).first()
+    if not bidder:
+        raise HTTPException(status_code=status.HTTP_406_NOT_ACCEPTABLE)
+    if bidder.is_canceled:
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT)
+    bidder.is_canceled = True
+    bidder.point += premire_point
+    user.point -= premire_point
+    db.commit()
+    return slot_response(slot)
 
-    exp_assignees = filter(lambda x: slot.task in x.exp_tasks, slot.assignees)
-    if len(slot.assignees) + 1 > slot.task.max_worker_num:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
-    if (slot.task not in user.exp_tasks) and slot.task.max_worker_num - len(
-        slot.assignees
-    ) + len(exp_assignees) <= slot.task.exp_worker_num:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
+def reassign(slot_id: str, user_id: str, db: Session):
+    slot = db.get(Slot, slot_id)
+    bid = slot.bid
+    if bid is None:
+        raise HTTPException(status_code=status.HTTP_405_METHOD_NOT_ALLOWED)
+    bidder = db.scalars(
+        select(Bidder)
+        .filter(Bidder.bid_id == bid.id, Bidder.user_id == user_id)
+        .limit(1)
+    ).first()
+    if not bidder:
+        raise HTTPException(status_code=status.HTTP_406_NOT_ACCEPTABLE)
+    bidder.is_canceled = False
+    db.commit()
+    user = db.get(User, user_id)
     slot.assignees.append(user)
     db.commit()
-    db.refresh(slot)
     return slot
 
 
-def complete(group_id, slot_id: str, done: bool, user: User, db: Session):
+def complete(slot_id: str,done:bool, user: User, db: Session):
     slot = db.get(Slot, slot_id)
     if slot.start_time > datetime.datetime.now():
         raise HTTPException(status_code=status.HTTP_406_NOT_ACCEPTABLE)
     slots = user.slots
     if slot not in slots:
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
-    group_user = db.scalars(
-        select(GroupUser).filter(
-            GroupUser.group_id == group_id, GroupUser.user_id == user.id
-        )
+    bid = slot.bid
+    bidder = db.scalars(
+        select(Bidder)
+        .filter(Bidder.bid_id == bid.id, Bidder.user_id == user.id)
+        .limit(1)
     ).first()
-    if not group_user:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
-
     slot.assignees.remove(user)
     if done:
-        user.exp_tasks.append(slot.task)
-        group_user.point += slot.task.point
+        user.exp_task.append(slot.task)
+        user.point += bidder.point
     db.commit()
-    db.refresh(slot)
-    return slot
+    return user_response(user)
 
 
-def bulk_delete(group_id:str,slots_id: list[str], db: Session):
-    delete_slot = []
+
+def bulk_delete(slots_id: list[str], db: Session):
     for slot_id in slots_id:
         slot = db.get(Slot, slot_id)
         if not slot:
@@ -158,24 +177,33 @@ def bulk_delete(group_id:str,slots_id: list[str], db: Session):
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail=f"id:{slot_id} is not found",
             )
-        if slot.task.group_id != group_id:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail="You don't have permission to delete this slot",
-            )
-        delete_slot.append({"id": slot.id, "name": slot.name})
+        if slot.template:
+            continue
         db.delete(slot)
     db.commit()
-    return delete_slot
+    return
 
 
-def delete_expired_slots(group_id: str, db: Session):
-    slots = db.scalars(
-        select(Slot).join(Slot.task).filter(Task.group_id == group_id)
-    ).all()
+def delete_prune_slots(db: Session):
+    prune_slots = []
+    slots = db.scalars(select(Slot)).all()
+    for slot in slots:
+        if slot.template == [] and slot.bid == None:
+            prune_slots.append({"id": slot.id, "name": slot.name})
+            db.delete(slot)
+    db.commit()
+    return prune_slots
+
+
+def delete_expired_slots(db: Session):
+    slots = db.scalars(select(Slot)).all()
     expired_slots = []
     for slot in slots:
-        if slot.end_time < datetime.datetime.now() and slot.assignees == []:
+        if (
+            slot.end_time < datetime.datetime.now()
+            and slot.assignees == []
+            and slot.template == []
+        ):
             expired_slots.append({"id": slot.id, "name": slot.name})
             db.delete(slot)
     db.commit()
